# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tsundoku API",
    "description": "Manga discovery service. Read endpoints (`GET`) are public unless `auth.read_requires_auth = true`. Write endpoints (`POST`) always require the admin bearer.",
    "contact": {
      "name": "Sylvain Cau",
      "email": "sylvain.cau@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/releases": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "List releases ordered by `observed_at` descending. Filters compose.",
        "operationId": "list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by resolution status (`resolved`, `unresolved`, `ambiguous`, `review_pending`).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceKind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seriesId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleasePage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/releases/unresolved": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "Review queue: releases awaiting human attention.",
        "description": "Returns releases whose status is `unresolved`, `ambiguous`, or\n`review_pending`, each with the recorded review candidates so the UI\ncan render a \"pick the right match\" panel without a second fetch.",
        "operationId": "list_unresolved",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "1-indexed page number.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Items per page (capped server-side at 200).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnresolvedPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/releases/{id}/link": {
      "post": {
        "tags": [
          "releases"
        ],
        "summary": "Manually link a release to a series. Body shape:",
        "description": "- `{ \"seriesId\": 42 }` — link to an existing series row by internal id.\n- `{ \"provider\": \"mangabaka\", \"externalId\": \"1677\" }` — link by a\n  provider's external id. If no `series_external_ids` row matches yet,\n  the active provider's `get` is called to fetch metadata and create\n  the series row before linking.",
        "operationId": "link",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Release id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseDto"
                }
              }
            }
          },
          "400": {
            "description": "Provider not registered or external_id unknown to provider"
          },
          "404": {
            "description": "Release or series not found"
          }
        },
        "security": [
          {
            "admin": []
          }
        ]
      }
    },
    "/api/v1/releases/{id}/reject": {
      "post": {
        "tags": [
          "releases"
        ],
        "summary": "Mark a release as \"not a series we care about\". Drops candidates and\npins the resolution status to `rejected` so the resolver leaves it\nalone on subsequent runs.",
        "operationId": "reject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Release id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseDto"
                }
              }
            }
          },
          "404": {
            "description": "Release not found"
          }
        },
        "security": [
          {
            "admin": []
          }
        ]
      }
    },
    "/api/v1/releases/{id}/retry": {
      "post": {
        "tags": [
          "releases"
        ],
        "summary": "Re-run the resolver against a single release. Useful after a provider\nrefresh, a config change, or a manual edit.",
        "operationId": "retry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Release id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseDto"
                }
              }
            }
          },
          "404": {
            "description": "Release not found"
          }
        },
        "security": [
          {
            "admin": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LinkRequest": {
        "type": "object",
        "description": "Body for the manual-link endpoint. Exactly one of:\n- `seriesId`: link to an existing series row by internal id.\n- `provider` + `externalId`: link via the named provider's external id;\n  the provider's `get` is called when no mapping exists yet.",
        "properties": {
          "externalId": {
            "type": [
              "string",
              "null"
            ]
          },
          "provider": {
            "type": [
              "string",
              "null"
            ]
          },
          "seriesId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "ReleaseDto": {
        "type": "object",
        "required": [
          "id",
          "sourceKind",
          "sourceName",
          "externalId",
          "title",
          "link",
          "files",
          "formats",
          "postedAt",
          "observedAt",
          "resolutionStatus",
          "resolutionAttempts"
        ],
        "properties": {
          "ddlUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "externalId": {
            "type": "string"
          },
          "files": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "formats": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "string"
          },
          "infoHash": {
            "type": [
              "string",
              "null"
            ]
          },
          "lastResolveAttemptAt": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "link": {
            "type": "string"
          },
          "magnet": {
            "type": [
              "string",
              "null"
            ]
          },
          "observedAt": {
            "type": "integer",
            "format": "int64"
          },
          "postedAt": {
            "type": "integer",
            "format": "int64"
          },
          "resolutionAttempts": {
            "type": "integer",
            "format": "int32"
          },
          "resolutionConfidence": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "resolutionPath": {
            "type": [
              "string",
              "null"
            ]
          },
          "resolutionStatus": {
            "type": "string"
          },
          "seriesId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "sizeBytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "sourceKind": {
            "type": "string"
          },
          "sourceName": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "torrentUrl": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ReleasePage": {
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "ReviewCandidateDto": {
        "type": "object",
        "required": [
          "seriesId",
          "seriesTitle",
          "score"
        ],
        "properties": {
          "reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "seriesCoverUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "seriesId": {
            "type": "integer",
            "format": "int32"
          },
          "seriesTitle": {
            "type": "string"
          }
        }
      },
...
      "UnresolvedPage": {
        "type": "object",
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UnresolvedRelease"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "UnresolvedRelease": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReleaseDto"
          },
          {
            "type": "object",
            "required": [
              "candidates"
            ],
            "properties": {
              "candidates": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ReviewCandidateDto"
                }
              }
            }
          }
        ]
      }
    },
    "securitySchemes": {
      "admin": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Admin bearer token (auth.admin_token in config)."
      }
    }
  },
...
}